Enable keyboard navigation on the close, and maximize button.
In order to enable keyboard navigation on the close, and maximize
buttons, set 'tabIndex' attribute, and 'onKeyDown' event for them.

Also - change the CSS for the 'tablist' to highlight the currently
focused 'tablist', which gives better user experience. It gives
indications to the user, which tab will be selected when the shortcut
keys (e.g. UP/LEFT/DOWN/RIGHT) for navigation.