Set the CSS 'display: none' for the element 'div[role=presentation]', having not children.
Element 'div[role=presentation]' is being created by 'rc-tabs', and it
does not have attribute 'tabIndex=0'. And - when keyboard is used for
navigation, it would get the focus.

And - it results into a confusion, as that element is not visible on the
screen, and user could not understand, which element is having focus at
the moment.

Hiding these elements (having no children), will give better user
experience while using keyboard for navigation.